enabling discount buy
