+++ pages/nav-config.tsx
@@ -1,4 +1,4 @@
-import { Coin, File, Gauge, LockLaminated } from 'phosphor-react';
+import { Coin, File, Gauge, Intersect, LockLaminated } from 'phosphor-react';
 export const navigation = [
   { name: 'Dashboard', href: '/', icon: <Gauge size={26} className="mr-2 w-8" /> },
   // {
@@ -12,17 +12,17 @@ export const navigation = [
   //     },
   //   ],
   // },
-  // {
-  //   name: 'Pre-Market Sale',
-  //   href: '/discount-buy',
-  //   icon: 'intersect',
-  //   subNav: [
-  //     {
-  //       name: 'Your Purchases',
-  //       href: '/discount-buy/your-purchases',
-  //     },
-  //   ],
-  // },
+  {
+    name: 'Discount Buy',
+    href: '/discount-buy',
+    icon: <Intersect size={26} className="mr-2 w-8" />,
+    subNav: [
+      {
+        name: 'Your Purchases',
+        href: '/discount-buy/your-purchases',
+      },
+    ],
+  },
   {
     name: 'Claim Tokens',
     href: '/claim',
